Week 2: Added some tests

+++ unit-tests/src/test/java/com/sap/cloud/s4hana/examples/addressmgr/commands/DeleteAddressCommandTest.java
@@ -49,7 +49,7 @@ public class DeleteAddressCommandTest {
 
     Integer result = new DeleteAddressCommand(service, "pbId", "addressId").execute();
 
-    Assert.assertEquals(new Integer(1), result);
+    Assert.assertEquals(new Integer(200), result);
   }
 
 }
